+++ draft-packages/table/CHANGELOG.md
@@ -3,6 +3,44 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [3.2.39](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/draft-table@3.2.38...@kaizen/draft-table@3.2.39) (2021-02-26)
+
+**Note:** Version bump only for package @kaizen/draft-table
+
+
+
+
+
+## [3.2.38](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/draft-table@3.2.37...@kaizen/draft-table@3.2.38) (2021-02-25)
+
+**Note:** Version bump only for package @kaizen/draft-table
+
+
+
+
+
+## [3.2.37](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/draft-table@3.2.36...@kaizen/draft-table@3.2.37) (2021-02-25)
+
+
+### Bug Fixes
+
+* Update minimum design-token dependency ([#1131](https://github.com/cultureamp/kaizen-design-system/issues/1131)) ([ce8182c](https://github.com/cultureamp/kaizen-design-system/commit/ce8182c054c9e8bc96bfdba8457bcd169d449204))
+
+
+
+
+
+## [3.2.36](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/draft-table@3.2.35...@kaizen/draft-table@3.2.36) (2021-02-25)
+
+
+### Bug Fixes
+
+* Use rgb-param variables in rgba ([#1094](https://github.com/cultureamp/kaizen-design-system/issues/1094)) ([4e7f0c7](https://github.com/cultureamp/kaizen-design-system/commit/4e7f0c7cbdadd5a0d606b58ed4b0f1344b8b9d99))
+
+
+
+
+
 ## [3.2.35](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/draft-table@3.2.34...@kaizen/draft-table@3.2.35) (2021-02-22)
 
 **Note:** Version bump only for package @kaizen/draft-table