+++ draft-packages/tile/CHANGELOG.md
@@ -3,6 +3,52 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [3.2.24](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/draft-tile@3.2.23...@kaizen/draft-tile@3.2.24) (2021-02-26)
+
+**Note:** Version bump only for package @kaizen/draft-tile
+
+
+
+
+
+## [3.2.23](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/draft-tile@3.2.22...@kaizen/draft-tile@3.2.23) (2021-02-25)
+
+
+### Bug Fixes
+
+* Update minimum design-token dependency ([#1131](https://github.com/cultureamp/kaizen-design-system/issues/1131)) ([ce8182c](https://github.com/cultureamp/kaizen-design-system/commit/ce8182c054c9e8bc96bfdba8457bcd169d449204))
+
+
+
+
+
+## [3.2.22](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/draft-tile@3.2.21...@kaizen/draft-tile@3.2.22) (2021-02-25)
+
+**Note:** Version bump only for package @kaizen/draft-tile
+
+
+
+
+
+## [3.2.21](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/draft-tile@3.2.20...@kaizen/draft-tile@3.2.21) (2021-02-25)
+
+**Note:** Version bump only for package @kaizen/draft-tile
+
+
+
+
+
+## [3.2.20](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/draft-tile@3.2.19...@kaizen/draft-tile@3.2.20) (2021-02-25)
+
+
+### Bug Fixes
+
+* Use rgb-param variables in rgba ([#1094](https://github.com/cultureamp/kaizen-design-system/issues/1094)) ([4e7f0c7](https://github.com/cultureamp/kaizen-design-system/commit/4e7f0c7cbdadd5a0d606b58ed4b0f1344b8b9d99))
+
+
+
+
+
 ## [3.2.19](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/draft-tile@3.2.18...@kaizen/draft-tile@3.2.19) (2021-02-22)
 
 **Note:** Version bump only for package @kaizen/draft-tile