+++ draft-packages/tooltip/CHANGELOG.md
@@ -3,6 +3,44 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.5.1](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/draft-tooltip@2.5.0...@kaizen/draft-tooltip@2.5.1) (2021-02-26)
+
+**Note:** Version bump only for package @kaizen/draft-tooltip
+
+
+
+
+
+# [2.5.0](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/draft-tooltip@2.4.4...@kaizen/draft-tooltip@2.5.0) (2021-02-25)
+
+
+### Features
+
+* Add automatic positioning for the Menu component, using popper ([#1110](https://github.com/cultureamp/kaizen-design-system/issues/1110)) ([aa1f289](https://github.com/cultureamp/kaizen-design-system/commit/aa1f2892364294fb5d959e3a57d3af910ac3f88f))
+
+
+
+
+
+## [2.4.4](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/draft-tooltip@2.4.3...@kaizen/draft-tooltip@2.4.4) (2021-02-25)
+
+
+### Bug Fixes
+
+* Update minimum design-token dependency ([#1131](https://github.com/cultureamp/kaizen-design-system/issues/1131)) ([ce8182c](https://github.com/cultureamp/kaizen-design-system/commit/ce8182c054c9e8bc96bfdba8457bcd169d449204))
+
+
+
+
+
+## [2.4.3](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/draft-tooltip@2.4.2...@kaizen/draft-tooltip@2.4.3) (2021-02-25)
+
+**Note:** Version bump only for package @kaizen/draft-tooltip
+
+
+
+
+
 ## [2.4.2](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/draft-tooltip@2.4.1...@kaizen/draft-tooltip@2.4.2) (2021-02-22)
 
 **Note:** Version bump only for package @kaizen/draft-tooltip